Modified - refactor home screen and splash screen
