+++ lib/screens/navigation_bottom_screen.dart
@@ -8,7 +8,7 @@ import '../classes/strings.dart';
 import '../provider/app_theme.dart';
 import '../widgets/main_drawer.dart';
 import '../widgets/profile_view.dart';
-import 'home_screen.dart';
+import '../features/home_feature/presentation/home_screen.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class NavigationBottomScreen extends StatefulWidget {
